Fixed errors when building lib_jre with "-x objective-c++" flag (more stringent casts, reserved words). 	Change on 2015/01/05 by tball <tball@google.com> ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=83280488

+++ jre_emul/android/libcore/luni/src/main/java/libcore/io/Memory.java
@@ -301,7 +301,7 @@ public final class Memory {
       if (!srcBytes) {
         return;
       }
-      memmove(dstBytes + dstOffset, srcBytes + srcOffset, byteCount);
+      memmove(dstBytes + dstOffset, srcBytes + srcOffset, (size_t)byteCount);
     ]-*/;
 
     public static native byte peekByte(long address) /*-[
